Alt-Shift-8/9 changed to Ctrl-8/9 for win and linux
because a lot of users in those systems prefer Ctrl-Shift or Alt-Shift
for language toggling. Also, Ctrl-numbers are usuallu used to switch
tabs, and high numbers like 8 and 9 are less probably will be used this
way.

In MacOS, Ctrl-Shift-8/9 is still reasonable, because it is not
conflicting with workcpaces (Ctrl-numbers) and there is no user habbit
to toggle language witj Ctrl-Shift.